bpf: refactor: Make core bpf_test.go logic reusable
Prepare for the next patch by moving some of the logic from bpf_test.go
into helpers.go and exporting it as a set of reusable functions and
types:

* ForEachSuite allows you to iterate over all BPF unit test collections
  one by one, transparently handling file reads and BPF collection
  preprocessing.
* CollectionToSuite converts a CollectionSpec and Collection for a BPF
  test into a set of Test structs. It performs all the validation on the
  test program structure to ensure the existence and uniqueness of
  setup/pkg/check programs for each test.
* Test contains all the logic for executing a BPF unit test, including
  the execution of its setup, pktgen, and check programs.
