+++ bpf/tests/bpftest/bpf_test.go
@@ -7,18 +7,13 @@ package bpftests
 import (
 	"bufio"
 	"bytes"
-	"encoding/binary"
-	"encoding/hex"
 	"errors"
 	"flag"
 	"fmt"
 	"io"
-	"io/fs"
-	"maps"
 	"os"
 	"path"
 	"regexp"
-	"slices"
 	"strings"
 	"testing"
 	"time"
@@ -28,11 +23,7 @@ import (
 	"github.com/cilium/ebpf/perf"
 	"github.com/cilium/ebpf/rlimit"
 	"github.com/cilium/hive/hivetest"
-	"github.com/davecgh/go-spew/spew"
-	"github.com/vishvananda/netlink/nl"
 	"golang.org/x/tools/cover"
-	"google.golang.org/protobuf/encoding/protowire"
-	"google.golang.org/protobuf/proto"
 
 	"github.com/cilium/cilium/pkg/bpf"
 	"github.com/cilium/cilium/pkg/datapath/link"
@@ -72,10 +63,7 @@ func TestBPF(t *testing.T) {
 		t.Log(err)
 	}
 
-	entries, err := os.ReadDir(*testPath)
-	if err != nil {
-		t.Fatal("os readdir: ", err)
-	}
+	logger := hivetest.Logger(t)
 
 	var instrLog io.Writer
 	if *testInstrumentationLog != "" {
@@ -89,30 +77,23 @@ func TestBPF(t *testing.T) {
 		instrLog = buf
 		defer buf.Flush()
 	}
-
 	mergedProfiles := make([]*cover.Profile, 0)
 
-	for _, entry := range entries {
-		if entry.IsDir() {
-			continue
+	if err := ForEachSuite(logger, *testPath, func(elfFile string, suiteSpec *ebpf.CollectionSpec) {
+		if *testFilePrefix != "" && !strings.HasPrefix(elfFile, *testFilePrefix) {
+			return
 		}
 
-		if !strings.HasSuffix(entry.Name(), ".o") {
-			continue
-		}
-
-		if *testFilePrefix != "" && !strings.HasPrefix(entry.Name(), *testFilePrefix) {
-			continue
-		}
-
-		t.Run(entry.Name(), func(t *testing.T) {
-			profiles := loadAndRunSpec(t, entry, instrLog)
+		t.Run(elfFile, func(t *testing.T) {
+			profiles := loadAndRunSpec(t, elfFile, suiteSpec, instrLog)
 			for _, profile := range profiles {
 				if len(profile.Blocks) > 0 {
 					mergedProfiles = addProfile(mergedProfiles, profile)
 				}
 			}
 		})
+	}); err != nil {
+		t.Fatal(err)
 	}
 
 	if *testCoverageReport != "" {
@@ -135,15 +116,13 @@ func TestBPF(t *testing.T) {
 	}
 }
 
-func loadAndRunSpec(t *testing.T, entry fs.DirEntry, instrLog io.Writer) []*cover.Profile {
+func loadAndRunSpec(t *testing.T, elfFile string, spec *ebpf.CollectionSpec, instrLog io.Writer) []*cover.Profile {
 	logger := hivetest.Logger(t)
-	elfPath := path.Join(*testPath, entry.Name())
+	elfPath := path.Join(*testPath, elfFile)
 
 	if instrLog != nil {
 		fmt.Fprintln(instrLog, "===", elfPath, "===")
 	}
-
-	spec := loadAndPrepSpec(t, elfPath)
 
 	var (
 		coll            *ebpf.Collection
@@ -154,13 +133,13 @@ func loadAndRunSpec(t *testing.T, entry fs.DirEntry, instrLog io.Writer) []*cove
 
 	if *testCoverageReport != "" {
 		if *noTestCoverage != "" {
-			matched, err := regexp.MatchString(*noTestCoverage, entry.Name())
+			matched, err := regexp.MatchString(*noTestCoverage, elfFile)
 			if err != nil {
 				t.Fatal("test file regex matching failed:", err)
 			}
 
 			if matched {
-				t.Logf("Disabling coverage report for %s", entry.Name())
+				t.Logf("Disabling coverage report for %s", elfFile)
 			}
 
 			collectCoverage = !matched
@@ -184,49 +163,9 @@ func loadAndRunSpec(t *testing.T, entry fs.DirEntry, instrLog io.Writer) []*cove
 	}
 	defer coll.Close()
 
-	testNameToPrograms := make(map[string]programSet)
-
-	checkProgUnique := func(prog *ebpf.Program, testName, progType string) {
-		if prog != nil {
-			t.Fatalf(
-				"File '%s' contains duplicate %s programs for the '%s' test.",
-				elfPath,
-				progType,
-				testName,
-			)
-		}
-	}
-
-	for progName, spec := range spec.Programs {
-		match := checkProgRegex.FindStringSubmatch(spec.SectionName)
-		if len(match) == 0 {
-			continue
-		}
-
-		progs := testNameToPrograms[match[1]]
-		if match[2] == "pktgen" {
-			checkProgUnique(progs.pktgenProg, match[1], match[2])
-			progs.pktgenProg = coll.Programs[progName]
-		}
-		if match[2] == "setup" {
-			checkProgUnique(progs.setupProg, match[1], match[2])
-			progs.setupProg = coll.Programs[progName]
-		}
-		if match[2] == "check" {
-			checkProgUnique(progs.checkProg, match[1], match[2])
-			progs.checkProg = coll.Programs[progName]
-		}
-		testNameToPrograms[match[1]] = progs
-	}
-
-	for testName, progSet := range testNameToPrograms {
-		if progSet.checkProg == nil {
-			t.Fatalf(
-				"File '%s' does not contain a check program for the '%s' test.",
-				elfPath,
-				testName,
-			)
-		}
+	suite, err := CollectionToSuite(spec, coll)
+	if err != nil {
+		t.Fatalf("File '%s': %v", elfPath, err)
 	}
 
 	// Collect debug events and add them as logs of the main test
@@ -250,15 +189,14 @@ func loadAndRunSpec(t *testing.T, entry fs.DirEntry, instrLog io.Writer) []*cove
 		}()
 	}
 
-	// Make sure sub-tests are executed in alphabetic order, to make test results repeatable if programs rely on
-	// the order of execution.
-	testNames := slices.Sorted(maps.Keys(testNameToPrograms))
+	for _, test := range suite {
+		if !test.HasCheck() {
+			continue
+		}
 
-	// Get maps used for common mocking facilities
-	skbMdMap := coll.Maps[mockSkbMetaMap]
-
-	for _, name := range testNames {
-		t.Run(name, subTest(testNameToPrograms[name], coll.Maps[suiteResultMap], skbMdMap))
+		t.Run(test.Name(), func(t *testing.T) {
+			test.Check(t, *dumpCtx)
+		})
 	}
 
 	if globalLogReader != nil {
@@ -289,300 +227,4 @@ func loadAndRunSpec(t *testing.T, entry fs.DirEntry, instrLog io.Writer) []*cove
 	}
 
 	return profiles
-}
-
-func loadAndPrepSpec(t *testing.T, elfPath string) *ebpf.CollectionSpec {
-	logger := hivetest.Logger(t)
-	spec, err := bpf.LoadCollectionSpec(logger, elfPath)
-	if err != nil {
-		t.Fatalf("load spec %s: %v", elfPath, err)
-	}
-
-	for _, m := range spec.Maps {
-		m.Pinning = ebpf.PinNone
-	}
-
-	for n, p := range spec.Programs {
-		switch p.Type {
-		case ebpf.XDP, ebpf.SchedACT, ebpf.SchedCLS:
-			continue
-		}
-
-		t.Logf("Skipping program '%s' of type '%s': BPF_PROG_RUN not supported", p.Name, p.Type)
-		delete(spec.Programs, n)
-	}
-
-	return spec
-}
-
-type programSet struct {
-	pktgenProg *ebpf.Program
-	setupProg  *ebpf.Program
-	checkProg  *ebpf.Program
-}
-
-var checkProgRegex = regexp.MustCompile(`[^/]+/test/([^/]+)/((?:check)|(?:setup)|(?:pktgen))`)
-
-const (
-	ResultSuccess = 1
-
-	suiteResultMap = "suite_result_map"
-	mockSkbMetaMap = "mock_skb_meta_map"
-)
-
-func subTest(progSet programSet, resultMap *ebpf.Map, skbMdMap *ebpf.Map) func(t *testing.T) {
-	return func(t *testing.T) {
-		// create ctx with the max allowed size(4k - head room - tailroom)
-		data := make([]byte, 4096-256-320)
-
-		// ctx is only used for tc programs
-		// non-empty ctx passed to non-tc programs will cause error: invalid argument
-		ctx := make([]byte, 0)
-		if progSet.checkProg.Type() == ebpf.SchedCLS {
-			// sizeof(struct __sk_buff) < 256, let's make it 256
-			ctx = make([]byte, 256)
-		}
-
-		var (
-			statusCode uint32
-			err        error
-		)
-		if progSet.pktgenProg != nil {
-			if statusCode, data, ctx, err = runBpfProgram(progSet.pktgenProg, data, ctx); err != nil {
-				t.Fatalf("error while running pktgen prog: %s", err)
-			}
-
-			if *dumpCtx {
-				t.Log("Pktgen returned status: ")
-				t.Log(statusCode)
-				t.Log("data after pktgen: ")
-				t.Log(spew.Sdump(data))
-				t.Log("ctx after pktgen: ")
-				t.Log(spew.Sdump(ctx))
-			}
-		}
-
-		if progSet.setupProg != nil {
-			if statusCode, data, ctx, err = runBpfProgram(progSet.setupProg, data, ctx); err != nil {
-				t.Fatalf("error while running setup prog: %s", err)
-			}
-
-			if *dumpCtx {
-				t.Log("Setup returned status: ")
-				t.Log(statusCode)
-				t.Log("data after setup: ")
-				t.Log(spew.Sdump(data))
-				t.Log("ctx after setup: ")
-				t.Log(spew.Sdump(ctx))
-			}
-
-			status := make([]byte, 4)
-			nl.NativeEndian().PutUint32(status, statusCode)
-			data = append(status, data...)
-		}
-
-		// Run test, input a
-		if statusCode, data, ctx, err = runBpfProgram(progSet.checkProg, data, ctx); err != nil {
-			t.Fatal("error while running check program:", err)
-		}
-
-		if *dumpCtx {
-			t.Log("Check returned status: ")
-			t.Log(statusCode)
-			t.Logf("data after check: %d", len(data))
-			t.Log(spew.Sdump(data))
-			t.Log("ctx after check: ")
-			t.Log(spew.Sdump(ctx))
-		}
-
-		// Clear map value after each test
-		defer func() {
-			for _, m := range []*ebpf.Map{resultMap, skbMdMap} {
-				if m == nil {
-					continue
-				}
-
-				var key int32
-				value := make([]byte, m.ValueSize())
-				m.Lookup(&key, &value)
-				for i := range value {
-					value[i] = 0
-				}
-				m.Update(&key, &value, ebpf.UpdateAny)
-			}
-		}()
-
-		var key int32
-		value := make([]byte, resultMap.ValueSize())
-		err = resultMap.Lookup(&key, &value)
-		if err != nil {
-			t.Fatal("error while getting suite result:", err)
-		}
-
-		// Detect the length of the result, since the proto.Unmarshal doesn't like trailing zeros.
-		valueLen := 0
-		valueC := value
-		for {
-			_, _, len := protowire.ConsumeField(valueC)
-			if len <= 0 {
-				break
-			}
-			valueLen += len
-			valueC = valueC[len:]
-		}
-
-		result := &SuiteResult{}
-		err = proto.Unmarshal(value[:valueLen], result)
-		if err != nil {
-			t.Fatal("error while unmarshalling suite result:", err)
-		}
-
-		for _, testResult := range result.Results {
-			// Remove the C-string, null-terminator.
-			name := strings.TrimSuffix(testResult.Name, "\x00")
-			t.Run(name, func(tt *testing.T) {
-				if len(testResult.TestLog) > 0 && testing.Verbose() || testResult.Status != SuiteResult_TestResult_PASS {
-					for _, log := range testResult.TestLog {
-						tt.Logf("%s", log.FmtString())
-					}
-				}
-
-				switch testResult.Status {
-				case SuiteResult_TestResult_ERROR:
-					tt.Fatal("Test failed due to unknown error in test framework")
-				case SuiteResult_TestResult_FAIL:
-					tt.Fail()
-				case SuiteResult_TestResult_SKIP:
-					tt.Skip()
-				}
-			})
-		}
-
-		if len(result.SuiteLog) > 0 && testing.Verbose() ||
-			SuiteResult_TestResult_TestStatus(statusCode) != SuiteResult_TestResult_PASS {
-			for _, log := range result.SuiteLog {
-				t.Logf("%s", log.FmtString())
-			}
-		}
-
-		switch SuiteResult_TestResult_TestStatus(statusCode) {
-		case SuiteResult_TestResult_ERROR:
-			t.Fatal("Test failed due to unknown error in test framework")
-		case SuiteResult_TestResult_FAIL:
-			t.Fail()
-		case SuiteResult_TestResult_SKIP:
-			t.SkipNow()
-		}
-	}
-}
-
-// A simplified version of fmt.Printf logic, the meaning of % specifiers changed to match the kernels printk specifiers.
-// In the eBPF code a user can for example call `test_log("expected 123, got %llu", some_val)` the %llu meaning
-// long-long-unsigned translates into a uint64, the rendered out would for example be -> 'expected 123, got 234'.
-// https://www.kernel.org/doc/Documentation/printk-formats.txt
-// https://github.com/libbpf/libbpf/blob/4eb6485c08867edaa5a0a81c64ddb23580420340/src/bpf_helper_defs.h#L152
-func (l *Log) FmtString() string {
-	var sb strings.Builder
-
-	end := len(l.Fmt)
-	argNum := 0
-
-	for i := 0; i < end; {
-		lasti := i
-		for i < end && l.Fmt[i] != '%' {
-			i++
-		}
-		if i > lasti {
-			sb.WriteString(strings.TrimSuffix(l.Fmt[lasti:i], "\x00"))
-		}
-		if i >= end {
-			// done processing format string
-			break
-		}
-
-		// Process one verb
-		i++
-
-		var spec []byte
-	loop:
-		for ; i < end; i++ {
-			c := l.Fmt[i]
-			switch c {
-			case 'd', 'i', 'u', 'x', 's':
-				spec = append(spec, c)
-				break loop
-			case 'l':
-				spec = append(spec, c)
-			default:
-				break loop
-			}
-		}
-		// Advance to to next char
-		i++
-
-		// No argument left over to print for the current verb.
-		if argNum >= len(l.Args) {
-			sb.WriteString("%!")
-			sb.WriteString(string(spec))
-			sb.WriteString("(MISSING)")
-			continue
-		}
-
-		switch string(spec) {
-		case "u":
-			fmt.Fprint(&sb, uint16(l.Args[argNum]))
-		case "d", "i", "s":
-			fmt.Fprint(&sb, int16(l.Args[argNum]))
-		case "x":
-			hb := make([]byte, 2)
-			binary.BigEndian.PutUint16(hb, uint16(l.Args[argNum]))
-			fmt.Fprint(&sb, hex.EncodeToString(hb))
-
-		case "lu":
-			fmt.Fprint(&sb, uint32(l.Args[argNum]))
-		case "ld", "li", "ls":
-			fmt.Fprint(&sb, int32(l.Args[argNum]))
-		case "lx":
-			hb := make([]byte, 4)
-			binary.BigEndian.PutUint32(hb, uint32(l.Args[argNum]))
-			fmt.Fprint(&sb, hex.EncodeToString(hb))
-
-		case "llu":
-			fmt.Fprint(&sb, uint64(l.Args[argNum]))
-		case "lld", "lli", "lls":
-			fmt.Fprint(&sb, int64(l.Args[argNum]))
-		case "llx":
-			hb := make([]byte, 8)
-			binary.BigEndian.PutUint64(hb, uint64(l.Args[argNum]))
-			fmt.Fprint(&sb, hex.EncodeToString(hb))
-
-		default:
-			sb.WriteString("%!")
-			sb.WriteString(string(spec))
-			sb.WriteString("(INVALID)")
-			continue
-		}
-
-		argNum++
-	}
-
-	return sb.String()
-}
-
-func runBpfProgram(prog *ebpf.Program, data, ctx []byte) (statusCode uint32, dataOut, ctxOut []byte, err error) {
-	dataOut = make([]byte, len(data))
-	if len(dataOut) > 0 {
-		// See comments at https://github.com/cilium/ebpf/blob/20c4d8896bdde990ce6b80d59a4262aa3ccb891d/prog.go#L563-L567
-		dataOut = make([]byte, len(data)+256+2)
-	}
-	ctxOut = make([]byte, len(ctx))
-	opts := &ebpf.RunOptions{
-		Data:       data,
-		DataOut:    dataOut,
-		Context:    ctx,
-		ContextOut: ctxOut,
-		Repeat:     1,
-	}
-	ret, err := prog.Run(opts)
-	return ret, opts.DataOut, ctxOut, err
 }
